add 'Microsoft Foundry (API Key)' connection type, update docs, add base anthropic client to use for native and foundry versions
