+++ Apps.Anthropic/Actions/BatchActions.cs
@@ -1,5 +1,4 @@
 ﻿using System.Text.RegularExpressions;
-using Apps.Anthropic.Api;
 using Apps.Anthropic.Constants;
 using Apps.Anthropic.Invocable;
 using Apps.Anthropic.Models.Entities;
@@ -16,6 +15,7 @@ using Blackbird.Xliff.Utils;
 using Blackbird.Xliff.Utils.Models;
 using Newtonsoft.Json;
 using RestSharp;
+using Apps.Anthropic.Api.Anthropic;
 
 namespace Apps.Anthropic.Actions;
