+++ Apps.Anthropic/Apps.Anthropic.csproj
@@ -5,7 +5,7 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <Product>Anthropic</Product>
-    <Version>1.3.8</Version>
+    <Version>1.3.9</Version>
     <Description>A next-generation AI assistant for your tasks, no matter the scale</Description>
     <AssemblyName>Apps.Anthropic</AssemblyName>
   </PropertyGroup>
