+++ Apps.Anthropic/Constants/ConnectionTypes.cs
@@ -5,6 +5,13 @@ public static class ConnectionTypes
     public const string AnthropicNative = "API Token";
     public const string BedrockCreds = "Amazon Bedrock (AWS Credentials)";
     public const string BedrockApiKey = "Amazon Bedrock (API Key)";
+    public const string MicrosoftFoundryApiKey = "MicrosoftFoundryApiKey";
 
-    public static readonly IEnumerable<string> SupportedConnectionTypes = [AnthropicNative, BedrockCreds, BedrockApiKey];
+    public static readonly IEnumerable<string> SupportedConnectionTypes = 
+    [
+        AnthropicNative, 
+        BedrockCreds, 
+        BedrockApiKey, 
+        MicrosoftFoundryApiKey
+    ];
 }
