+++ Apps.Anthropic/Constants/JsonOptions.cs
@@ -1,0 +1,11 @@
+﻿using Newtonsoft.Json;
+
+namespace Apps.Anthropic.Constants;
+
+public static class JsonOptions
+{
+    public static JsonSerializerSettings JsonSettings => new()
+    {
+        NullValueHandling = NullValueHandling.Ignore,
+    };
+}
